SINF-348 - tidy up some test classes
