+++ src/test/java/uk/gov/crowncommercial/dsd/api/auth/AbstractAccountApiTest.java
@@ -1,26 +1,10 @@
 package uk.gov.crowncommercial.dsd.api.auth;
 
-import org.apache.camel.CamelContext;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-import com.github.tomakehurst.wiremock.client.WireMock;
 import io.restassured.RestAssured;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-@DirtiesContext
-@AutoConfigureWireMock(port = 0)
-abstract class AbstractAccountApiTest {
-
-  static final String AUTH_BEARER_TOKEN = "Bearer abc123";
+abstract class AbstractAccountApiTest extends AbstractApiTest {
 
   @Value("${api.paths.base.catalog}")
   String apiCatalogBasePath;
@@ -28,22 +12,11 @@ abstract class AbstractAccountApiTest {
   @Value("${spree.api.paths.base.storefront}")
   String spreeApiCatalogBasePath;
 
-
-  @Autowired
-  CamelContext camelContext;
-
-  @LocalServerPort
-  int port;
-
+  @Override
   @BeforeEach
   public void beforeEach() {
-    RestAssured.port = port;
+    super.beforeEach();
     RestAssured.basePath = apiCatalogBasePath;
-  }
-
-  @AfterEach
-  public void afterEach() {
-    WireMock.reset();
   }
 
 }
