+++ src/test/java/uk/gov/crowncommercial/dsd/api/auth/AbstractOauthApiTest.java
@@ -1,26 +1,10 @@
 package uk.gov.crowncommercial.dsd.api.auth;
 
-import org.apache.camel.CamelContext;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-import com.github.tomakehurst.wiremock.client.WireMock;
 import io.restassured.RestAssured;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-@DirtiesContext
-@AutoConfigureWireMock(port = 0)
-abstract class AbstractOauthApiTest {
-
-  static final String AUTH_BEARER_TOKEN = "Bearer abc123";
+abstract class AbstractOauthApiTest extends AbstractApiTest {
 
   @Value("${api.paths.base.oauth}")
   String apiOauthBasePath;
@@ -28,21 +12,11 @@ abstract class AbstractOauthApiTest {
   @Value("${spree.api.paths.base.oauth}")
   String spreeApiOauthBasePath;
 
-  @Autowired
-  CamelContext camelContext;
-
-  @LocalServerPort
-  int port;
-
+  @Override
   @BeforeEach
   public void beforeEach() {
-    RestAssured.port = port;
+    super.beforeEach();
     RestAssured.basePath = apiOauthBasePath;
-  }
-
-  @AfterEach
-  public void afterEach() {
-    WireMock.reset();
   }
 
 }
